# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/timesheets:
    get:
      tags:
        - Timesheets
      summary: List Timesheets
      description: Get a paginated list of Timesheets with optional filtering and sorting
      operationId: listTimesheets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Timesheets
      summary: Create Timesheet
      description: Create a new Timesheet
      operationId: createTimesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimesheetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Timesheet:
      type: object
      required:
        - id
        - company_id
        - project_id
        - currency
        - status
        - total_hours
        - total_billable_amount
        - total_costing_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        employee_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        status:
          type: string
        total_hours:
          type: string
        total_billable_amount:
          type: string
        total_costing_amount:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateTimesheetRequest:
      type: object
      required:
        - company_id
        - project_id
        - currency
        - status
        - total_hours
        - total_billable_amount
        - total_costing_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        employee_id:
          nullable: true
          type: string
          format: uuid
        currency:
          type: string
        status:
          type: string
        total_hours:
          type: string
        total_billable_amount:
          type: string
        total_costing_amount:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    TimesheetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Timesheet'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []